youtube-comments-api/requirements.txt
flask==2.0.1
google-api-python-client==2.33.0
pandas==1.3.3
python-dotenv==0.19.0